Pre-process event if it comes from an SNS topic
AWS only allows one action trigger for each event type per object.
If multiple triggers must be configured, the usual way to solve this
is to use a fan-out architecture using AWS SNS.

When an S3 event occurs, it's published to an SNS topic. The multiple
lambdas can then be added as subscribers to that topic. This allows us
to trigger multiple lambdas in parallel from a single S3 event.